refactor: move HomeItemsType from domain to UI
