+++ app/src/main/res/layout/item_airing_today.xml
@@ -5,7 +5,7 @@
 
     <data>
 
-        <import type="com.karrar.movieapp.domain.enums.HomeItemsType" />
+        <import type="com.karrar.movieapp.ui.home.homeUiState.HomeItemsType" />
 
         <variable
             name="listener"
