+++ app/src/main/res/layout/item_similar_movie.xml
@@ -15,7 +15,7 @@
 
         <variable
             name="type"
-            type="com.karrar.movieapp.domain.enums.HomeItemsType" />
+            type="com.karrar.movieapp.ui.home.homeUiState.HomeItemsType" />
 
     </data>
