+++ app/src/main/res/layout/item_tvshow.xml
@@ -14,7 +14,7 @@
 
         <variable
             name="type"
-            type="com.karrar.movieapp.domain.enums.HomeItemsType" />
+            type="com.karrar.movieapp.ui.home.homeUiState.HomeItemsType" />
 
     </data>
