+++ app/src/main/res/layout/list_movie.xml
@@ -11,7 +11,7 @@
 
         <variable
             name="movieType"
-            type="com.karrar.movieapp.domain.enums.HomeItemsType" />
+            type="com.karrar.movieapp.ui.home.homeUiState.HomeItemsType" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
